requirements.txt
Django==3.2.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
pytz==2021.1
